harden(hpke): RFC HKDF-expand cap, u64 sequence counter, zeroize auxiliary KEM secrets [skip ci]
Wave 1 of the fork-audit review backlog (hardening; the live path was already
sound, so these are defense-in-depth/robustness, not vuln fixes):

- kdf/hkdf.rs: enforce the RFC 5869 / RFC 9180 L <= 255*Nh bound in the
  low-level expand/expand_static (previously only the export() wrapper capped).
- hpke_session/hpke_core/wasm: widen the AEAD sequence counter u32 -> u64
  (compute_nonce already I2OSP'd via u64, so nonce bytes are unchanged for
  existing vectors); raises the per-context message ceiling. It already errored
  at the limit, so this is not a nonce-reuse fix.
- kem (auxiliary Kem trait / MlKemImpl): return secret keys and shared secrets
  in Zeroizing<Vec<u8>>. The live HPKE path (KemProvider) already used
  SecretBytes; this hardens the public-but-non-core abstraction.

hpke clippy (--all-targets --all-features -D warnings) and full test suite pass;
workspace clippy clean.

+++ lib-q-hpke/src/hpke_session.rs
@@ -34,9 +34,9 @@ pub struct HpkeSenderContext {
     /// Encapsulated key to be sent to receiver
     pub encapsulated_key: Vec<u8>,
     /// Sequence number
-    pub sequence_number: u32,
+    pub sequence_number: u64,
     /// Maximum sequence number before context must be rekeyed
-    pub max_sequence_number: u32,
+    pub max_sequence_number: u64,
     /// Context state
     pub state: HpkeContextState,
     /// Cryptographic backend for multi-message seal and export
@@ -83,7 +83,7 @@ impl HpkeSenderContext {
             aead,
             encapsulated_key,
             sequence_number: 0,
-            max_sequence_number: u32::MAX - 1,
+            max_sequence_number: u64::MAX - 1,
             state: HpkeContextState::Active,
             hpke_crypto,
         }
@@ -139,9 +139,9 @@ pub struct HpkeReceiverContext {
     /// AEAD algorithm from the negotiated cipher suite
     pub aead: HpkeAead,
     /// Sequence number
-    pub sequence_number: u32,
+    pub sequence_number: u64,
     /// Maximum sequence number before context must be rekeyed
-    pub max_sequence_number: u32,
+    pub max_sequence_number: u64,
     /// Context state
     pub state: HpkeContextState,
     /// Cryptographic backend for multi-message open and export
@@ -184,7 +184,7 @@ impl HpkeReceiverContext {
             cipher_suite,
             aead,
             sequence_number: 0,
-            max_sequence_number: u32::MAX - 1,
+            max_sequence_number: u64::MAX - 1,
             state: HpkeContextState::Active,
             hpke_crypto,
         }
